Separate model and ui in two submodules
