+++ netlogo_pyqt5/agents/__init__.py
@@ -1,0 +1,2 @@
+from agents.model import *
+from agents.ui import *
